Refactor duplicate __sleep and __wakeup methods
